+++ docs/changelog.rst
@@ -21,6 +21,7 @@ Changelog
 * :feature:`10868` Farcaster Pro Purchases will now be properly understood by rotki.
 * :feature:`10712` bitcoin events can now be filtered by address.
 * :feature:`10542` StakeDAO v2 votemarket events will now be properly decoded.
+* :feature:`12030` Historical prices from cache are now filtered to your enabled oracles.
 * :bug:`11976` Bybit deposits and withdrawals older than 30 days will now be correctly imported.
 * :bug:`11964` Resolving many internal transaction conflicts at once will no longer flood the frontend with duplicate refresh requests, reducing unnecessary network calls and improving responsiveness.
 * :bug:`-` rotki no longer gets stuck when querying Routescan transactions with more than 100 internal transactions.